+++ components/config/src/lib.rs
@@ -1,7 +1,9 @@
 mod config;
 pub mod highlighting;
 mod theme;
-pub use crate::config::{Config, languages::Language, link_checker::LinkChecker, taxonomies::Taxonomy};
+pub use crate::config::{
+    languages::Language, link_checker::LinkChecker, taxonomies::Taxonomy, Config,
+};
 
 use std::path::Path;