+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix parsing of dates in arrays in `extra`
 - Add a `--force` argument to `zola init` to allow creating a Zola site in a non-empty directory
 - Make themes more flexible: `include` can now be used
+- Make search index generation configurable
 
 ## 0.11.0 (2020-05-25)